refactor(linter): Move miscellaneous linters to new file
