+++ tasks/linter/__init__.py
@@ -1,4 +1,5 @@
 """Module regrouping all invoke tasks used for linting the `datadog-agent` repo"""
 
+from .misc import copyrights, filenames  # noqa: F401
 from .old import *  # noqa: F403
 from .python import python  # noqa: F401
